🏗️ Architect: Modularized Focuser God Object
Refactored the 854-line `apts/opticalequipment/focuser.py` into a modular package structure.
- Isolated core `Focuser` logic in `apts/opticalequipment/focuser/base.py`.
- Distributed 47 database entries and factory methods across 23 vendor-specific modules in `apts/opticalequipment/focuser/vendors/`.
- Implemented a dynamic aggregation mechanism in `apts/opticalequipment/focuser/__init__.py` to maintain the public API.
- Verified with 660 unit tests and confirmed backward compatibility for all imports.
