+++ apts/opticalequipment/focuser/vendors/celestron.py
@@ -1,0 +1,11 @@
+from ..base import Focuser
+
+
+class CelestronFocuser(Focuser):
+    _DATABASE = {
+        "Celestron_2_Crayford_Focuser": {'brand': 'Celestron', 'name': '2" Crayford Focuser', 'type': 'type_focuser', 'optical_length': 0, 'mass': 500, 'tside_thread': '', 'tside_gender': '', 'cside_thread': '', 'cside_gender': '', 'reversible': False, 'bf_role': ''},
+    }
+
+    @classmethod
+    def Celestron_2_Crayford_Focuser(cls):
+        return cls.from_database(cls._DATABASE["Celestron_2_Crayford_Focuser"])
