+++ apts/opticalequipment/focuser/vendors/qhy.py
@@ -1,0 +1,11 @@
+from ..base import Focuser
+
+
+class QHYFocuser(Focuser):
+    _DATABASE = {
+        "QHY_Q_Focuser": {'brand': 'QHY', 'name': 'Q-Focuser', 'type': 'type_focuser', 'optical_length': 0, 'mass': 180, 'tside_thread': '', 'tside_gender': '', 'cside_thread': '', 'cside_gender': '', 'reversible': False, 'bf_role': ''},
+    }
+
+    @classmethod
+    def QHY_Q_Focuser(cls):
+        return cls.from_database(cls._DATABASE["QHY_Q_Focuser"])
